Update main.py
Deleted the kv language from here and put it in another file. If I've screwed up this is easy to fix. I believe Builder.load_string() can be removed but I'm not sure. Please test the app to make sure it behaves as before, as I am having trouble with the mac version of kivy.