Kivy Settings
I’ve changed the “settings” button to go to the settings page. However,
I have not added a transition.

Main.py now imports the settings and sets defaults. settingsjson.py is
the necessary json for the settings to work.

my.ini is a file for easy readability of the current settings. It is
automatically updated.