Organised Navigation
This adds the ability to create a new button on the sidebar. I have
added some other things that don’t work but I must commit now and
they’ll be fixed in the next commit. My next step will be to add a
feature to be able to name these notebooks, and then for them to
actually link to something.